gcd: speed up startup by avoiding "uniq"
Overlapping patterns can result in entries that "uniq" cannot filter
because they are not output consecutively. We also don't
want to use "sort -u" because it results in a noticeable delay.

Speed things up by avoiding "uniq" altogether. fzf does a fine job
as-is even when the output contains a few repeated entries.